% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr.R
\name{fr_agencies}
\alias{fr_agencies}
\title{List Federal Register agencies}
\usage{
fr_agencies()
}
\value{
tibble: id, name, short_name, slug, url, parent_id
}
\description{
Returns all federal agencies that publish in the Federal Register.
}
