% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr.R
\name{fr_document}
\alias{fr_document}
\title{Fetch a single Federal Register document}
\usage{
fr_document(document_number)
}
\arguments{
\item{document_number}{Document number (e.g. "2026-06029")}
}
\value{
tibble: one row with document details
}
\description{
Fetch a single Federal Register document
}
